+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+<h4>0.18.0</h4>
+
+* Update UserAPI class to use new endpoint
+* Added Workflow Classes: workflow, stage, task
+
 <h4>0.17.4</h4>
 
 * rubocop fixes